+++ lib/config/definitions.js
@@ -361,15 +361,16 @@ const options = [
     description: 'Version scheme to use for filtering and comparisons',
     type: 'string',
     allowedValues: [
+      'cargo',
+      'composer',
       'docker',
+      'hashicorp',
       'loose',
+      'npm',
       'pep440',
-      'semver',
-      'semver-cargo',
-      'semver-composer',
-      'semver-hashicorp',
+      'ruby',
     ],
-    default: 'semver',
+    default: 'npm',
     cli: false,
     env: false,
   },
@@ -1090,7 +1091,7 @@ const options = [
     default: {
       commitMessageTopic: 'Terraform module {{depNameShort}}',
       fileMatch: ['\\.tf$'],
-      versionScheme: 'semver-hashicorp',
+      versionScheme: 'hashicorp',
     },
     mergeable: true,
   },
@@ -1266,7 +1267,7 @@ const options = [
     type: 'json',
     default: {
       fileMatch: ['(^|/)([\\w-]*)composer.json$'],
-      versionScheme: 'semver-composer',
+      versionScheme: 'composer',
       ignorePath: ['**/vendor/**'],
     },
     mergeable: true,
