+++ lib/manager/composer/extract.js
@@ -1,5 +1,5 @@
 const is = require('@sindresorhus/is');
-const semverComposer = require('../../versioning/semver-composer');
+const semverComposer = require('../../versioning/composer');
 
 module.exports = {
   extractPackageFile,
