+++ lib/manager/gomod/extract.js
@@ -1,4 +1,4 @@
-const { isVersion } = require('../../versioning/semver');
+const { isVersion } = require('../../versioning/npm');
 
 module.exports = {
   extractPackageFile,
