+++ lib/manager/npm/extract/index.js
@@ -3,7 +3,7 @@ const upath = require('upath');
 const { getLockedVersions } = require('./locked-versions');
 const { detectMonorepos } = require('./monorepo');
 const { mightBeABrowserLibrary } = require('./type');
-const semver = require('../../../versioning/semver');
+const semver = require('../../../versioning/npm');
 
 module.exports = {
   extractAllPackageFiles,
