+++ lib/manager/nvm/extract.js
@@ -1,4 +1,4 @@
-const semver = require('../../versioning/semver');
+const semver = require('../../versioning/npm');
 
 module.exports = {
   extractPackageFile,
