+++ lib/manager/terraform/extract.js
@@ -1,4 +1,4 @@
-const { isValid, isVersion } = require('../../versioning/semver-hashicorp');
+const { isValid, isVersion } = require('../../versioning/hashicorp');
 
 module.exports = {
   extractPackageFile,
