+++ lib/manager/travis/package.js
@@ -1,7 +1,7 @@
 const is = require('@sindresorhus/is');
 const { isEqual } = require('lodash');
 const { getPkgReleases } = require('../../datasource');
-const { isVersion, maxSatisfyingVersion } = require('../../versioning/semver');
+const { isVersion, maxSatisfyingVersion } = require('../../versioning/npm');
 
 module.exports = {
   getPackageUpdates,
