+++ lib/versioning/cargo/index.js
@@ -1,4 +1,4 @@
-const semver = require('../semver');
+const semver = require('../npm');
 
 // NOTE: Partial versions like '1.2' don't get converted to '^1.2'
 // because isVersion('1.2') === false
