+++ lib/versioning/composer/index.js
@@ -1,4 +1,4 @@
-const semver = require('../semver');
+const semver = require('../npm');
 
 function padZeroes(input) {
   const sections = input.split('.');
