+++ lib/versioning/index.js
@@ -15,12 +15,12 @@ module.exports = {
 function get(versionScheme) {
   if (!versionScheme) {
     logger.debug('Missing versionScheme');
-    return schemes.semver;
+    return schemes.npm;
   }
   const scheme = schemes[versionScheme];
   if (!scheme) {
     logger.warn({ versionScheme }, 'Unknown version scheme');
-    return schemes.semver;
+    return schemes.npm;
   }
   return scheme;
 }
