+++ lib/versioning/ruby/operator.js
