+++ lib/versioning/ruby/strategies/bump.js
