+++ lib/versioning/ruby/strategies/index.js
