+++ lib/versioning/ruby/strategies/pin.js
