+++ lib/versioning/ruby/strategies/replace.js
