+++ test/datasource/packagist.spec.js
@@ -20,7 +20,7 @@ describe('datasource/packagist', () => {
       hostRules.find = jest.fn(input => input);
       global.repoCache = {};
       config = {
-        versionScheme: 'semver-composer',
+        versionScheme: 'composer',
         registryUrls: [
           {
             type: 'composer',
